feat: 예외 처리

+++ src/main/java/EBus/EBusback/domain/notice/service/NoticeService.java
@@ -5,6 +5,7 @@ import java.util.stream.Collectors;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.server.ResponseStatusException;
 
 import EBus.EBusback.domain.member.entity.Member;
 import EBus.EBusback.domain.member.entity.Role;
@@ -13,6 +14,7 @@ import EBus.EBusback.domain.notice.entity.Notice;
 import EBus.EBusback.domain.notice.repository.NoticeRepository;
 import EBus.EBusback.domain.post.dto.PostRequestDto;
 import EBus.EBusback.global.SecurityUtil;
+import EBus.EBusback.global.exception.ErrorCode;
 import lombok.RequiredArgsConstructor;
 
 @Service
@@ -24,8 +26,10 @@ public class NoticeService {
 
 	public NoticeResponseDto createNotice(PostRequestDto requestDto) {
 		Member member = SecurityUtil.getCurrentUser();
+		if (member == null)
+			throw new ResponseStatusException(ErrorCode.NON_LOGIN.getStatus(), ErrorCode.NON_LOGIN.getMessage());
 		if (!member.getRole().equals(Role.ADMIN))
-			throw new RuntimeException("공지사항은 관리자만 작성 가능합니다.");
+			throw new ResponseStatusException(ErrorCode.NO_ADMIN.getStatus(), ErrorCode.NO_ADMIN.getMessage());
 		return new NoticeResponseDto(noticeRepository.save(
 			Notice.builder().title(requestDto.getTitle()).content(requestDto.getContent()).writer(member).build()
 		));
@@ -33,7 +37,8 @@ public class NoticeService {
 
 	public NoticeResponseDto findNotice(Long noticeId) {
 		Notice notice = noticeRepository.findById(noticeId)
-			.orElseThrow(() -> new RuntimeException("공지사항을 찾을 수 없습니다."));
+			.orElseThrow(() -> new ResponseStatusException(
+				ErrorCode.NO_NOTICE_EXIST.getStatus(), ErrorCode.NO_NOTICE_EXIST.getMessage()));
 		return new NoticeResponseDto(notice);
 	}
 
@@ -44,10 +49,13 @@ public class NoticeService {
 
 	public void removeNotice(Long noticeId) {
 		Member member = SecurityUtil.getCurrentUser();
+		if (member == null)
+			throw new ResponseStatusException(ErrorCode.NON_LOGIN.getStatus(), ErrorCode.NON_LOGIN.getMessage());
 		Notice notice = noticeRepository.findById(noticeId)
-			.orElseThrow(() -> new RuntimeException("공지사항을 찾을 수 없습니다."));
+			.orElseThrow(() -> new ResponseStatusException(
+				ErrorCode.NO_NOTICE_EXIST.getStatus(), ErrorCode.NO_NOTICE_EXIST.getMessage()));
 		if (!member.getRole().equals(Role.ADMIN))
-			throw new RuntimeException("관리자만 삭제할 수 있습니다.");
+			throw new ResponseStatusException(ErrorCode.NO_ADMIN.getStatus(), ErrorCode.NO_ADMIN.getMessage());
 		noticeRepository.delete(notice);
 	}
 }
